Make thrift transport configurable
Thrift transport can be specified on a per-connection basis by providing the
fully qualified name of an ITransportFactory implementation. This can be
set using a new 'transportFactory' querystring parameter on the connection
url, or directly on a CassandraDatasource using the setTransportFactory
method. Specifying the factory is optional with the default being to use
TFramedTransport.

The ITransportFactory interface was added to Cassandra in version 1.1.5
but in that version is not included in either the cassandra-thrift nor
cassandra-client-util jars so it and a default implementation,
TFramedTransportFactory, have been duplicated here. Ideally this duplication
will not be necessary in future versions.